Add switch to disable KParts' handling of window titles
Applications that use several KParts, like KDevelop, want to do the
window title settings themselves. To enable that without causing a
flickering of the KParts' title and the application's title, an
explicit switch is required, since simply removing the setWindowTitle
connection after executing a KParts::MainWindow::createGUI(...) call still
causes an initial flickering, when the GUIActivateEvent is sent in
the createGUI method.

REVIEW: 128249