KParts partviewer test app: add action list to switch parts
Summary:
This is to show how an application would use partsForMimeType() in practice.
The other obvious usage is for keditfiletype itself, which only needs the list, and not to
actually load the part (it then lets the user reorder and it writes out into mimeapps.list).

Test Plan: bin/partviewer /tmp/foo.png, then look at the File menu for actions to switch parts.

Reviewers: aacid

Differential Revision: https://phabricator.kde.org/D27974