Deprecate KParts::Plugin class
This is only used in Konqueror, there the class is imported & the plugins are ported to
the imported class. See https://invent.kde.org/network/konqueror/-/merge_requests/99.

Also the plugins need to be ported to embedded json metadata, which is more easily doable
in the imported class in Konqueror.

Task: https://phabricator.kde.org/T13806